Add types, declare modules
