+++ globals.d.ts
@@ -1,8 +1,5 @@
 declare module 'bn.js';
 declare module 'ethereumjs-abi';
-declare module 'es6-promisify';
-declare module 'yargs';
-declare module 'solc';
 
 declare module '*.json' {
     const json: any;
@@ -20,30 +17,16 @@ declare var beforeEach: any;
 declare var describe: any;
 declare var it: any;
 
-declare module 'ethereumjs-util' {
-  function bufferToHex(value: Buffer): string;
-  function ecrecover(msgHash: Buffer, v: number, r: Buffer, s: Buffer): Buffer;
-  function fromRpcSig(sig: string): {v: number, r: Buffer, s: Buffer};
-  function hashPersonalMessage(hash: Buffer): Buffer;
-  function isHexString(value: any): boolean;
-  function pubToAddress(pubKey: Buffer, sanitize?: boolean): Buffer;
-  function setLength(a: Buffer, length: number): Buffer;
-  function setLengthLeft(a: Buffer, length: number): Buffer;
-  function sha3(a: Buffer|string|number, bits?: number): Buffer;
-  function toBuffer(value: any): Buffer;
-  function isValidAddress(address: string): boolean;
+declare module 'solc' {
+    function compile(sources: any, optimizerEnabled: number, findImports: (importPath: string) => any): any;
+    function setupMethods(solcBin: any): any;
+    export = {
+        compile,
+        setupMethods,
+    };
+}
 
-  export = {
-            bufferToHex,
-            ecrecover,
-            fromRpcSig,
-            hashPersonalMessage,
-            isHexString,
-            pubToAddress,
-            setLength,
-            setLengthLeft,
-            sha3,
-            toBuffer,
-            isValidAddress,
-          };
+declare module 'es6-promisify' {
+    function promisify(original: any, settings?: any): ((...arg: any[]) => Promise<any>);
+    export = promisify;
 }
