--- scripts/schemas/compiler_schema.ts
@@ -1,10 +1,0 @@
-export interface CompilerOptions {
-    contractsDir: string;
-    networkId: number;
-    optimizerRuns: number;
-    artifactsDir: string;
-}
-
-export interface ContractSources {
-    [key: string]: string;
-}
