+++ util/types.ts
@@ -1,4 +1,5 @@
 import * as BigNumber from 'bignumber.js';
+import * as Web3 from 'web3';
 
 export interface BalancesByOwner {
   [ownerAddress: string]: {
@@ -131,3 +132,41 @@ export enum ExchangeContractErrs {
   ERROR_ROUNDING_ERROR_TOO_LARGE,
   ERROR_INSUFFICIENT_BALANCE_OR_ALLOWANCE,
 }
+
+export interface ContractArtifact {
+    contract_name: string;
+    networks: ContractNetworks;
+};
+
+export interface ContractNetworks {
+    [key: number]: ContractData;
+};
+
+export interface ContractData {
+    solc_version: string;
+    optimizer_runs: number;
+    keccak256: string;
+    abi: Web3.ContractAbi;
+    unlinked_binary: string;
+    address?: string;
+    updated_at: number;
+};
+
+export interface SolcErrors {
+    [key: string]: boolean;
+};
+
+export interface CompilerOptions {
+    contractsDir: string;
+    networkId: number;
+    optimizerRuns: number;
+    artifactsDir: string;
+};
+
+export interface ContractSources {
+    [key: string]: string;
+};
+
+export interface ImportContents {
+    contents: string;
+};
